// src/api/interface/menuIterface.ts
interface OrderItem {
  created_at: string // ISO 8601 date string
  description: string // Description of the product
  extra_toppings: string // Additional toppings selected
  id: number // Unique identifier for the order item
  instruction: string // Special instructions for the order
  name: string | null // Name of the product, nullable
  net_price: string // Net price as a string
  order: number // Order ID this item belongs to
  product_sku: string // SKU of the product
  quantity: number // Quantity of the item
  size: string // Size of the product
  spice_level: string // Spice level selected
  status: string // Current status of the order item
  tax_price: string // Tax amount as a string
  total_price: string // Total price as a string
  updated_at: string // ISO 8601 date string for last update
  resturant: number
}

export interface SearchedMenuItem {
  id: number
  sku: string
  title: string
  status: 'pending' | 'active' | 'inactive' // Assuming possible statuses
  description: string
  availability: 'in stock' | 'out of stock' | 'preorder' // Adjust as needed
  condition: 'New' | 'Used' | 'Refurbished' | string
  price: number // Changed to number for price representation
  link: string
  image_link: string
  brand: string
  updated_at: string // ISO date string
  created_at: string // ISO date string
  google_product_category?: string
  fb_product_category?: string
  quantity_to_sell_on_facebook: number
  sale_price?: number // Assuming sale_price is numeric
  sale_price_effective_date?: string // ISO date range or specific format
  item_group_id?: string
  gender: 'male' | 'female' | 'unisex' | string
  color?: string
  size?: string
  age_group: 'newborn' | 'infant' | 'toddler' | 'children' | 'all ages' | string
  material?: string
  pattern?: string
  shipping?: string
  shipping_weight: string // Consider using a more specific type or separate fields
  video_url?: string
  address_city?: string
  address_country?: string
  address_neighborhoods?: string
  address_postal_code?: string
  address_region?: string
  address_street_address?: string
  availability_polygon_coordinates_latitude: number[]
  availability_polygon_coordinates_longitude: number[]
  availability_circle_origin_latitude: number
  availability_circle_origin_longitude: number
  availability_circle_radius_unit: 'KM' | 'MILES' | string
  availability_circle_radius: number
  style?: string
  menu_number: string
  business: number
  restaurant?: number | null // Assuming restaurant is a number or null
  quantity: number
  order_number: string
  total_price: string
  delivery_type: string
  order_items: OrderItem[]
}